Refactored FormsProvider
